remove parsing URL/req. method, ask instead
Instead of parsing the URL and request method from the buffer itself, we
now have to provide emacs with that information through
interactivity. The first time `es-query-region` is called the user
should be requested to input the URL and request method.

The reason for doing this is that we want to be able to use the queries
in other contexts besides emacs.