refactor requests into 1 interactive function
Having multiple requests functions is confusing. This patch refactors
the different ways into 1 do-what-I-mean function. If you run it inside
a request body that does not have a header, the url and method is taken
from the vars. If the request body has a header, that header is used
instead. If you provide the function with a prefix `C-u` it will send a
request for each of the request-bodies in the buffer.