Delay evaluation of defaults for org-babel
Documentation states that, when :url and :method are not provided to
"es" org-babel blocks, that "es-default-url" and
"es-default-request-method" will be used. However, without this
change, that does not appear to be the case.

Instead, it appears to be using old versions of these values from
"org-babel-default-header-args:es". Thus, changing the value of
"es-default-*" *will* affect es-mode, but *will not* affect "es"
org-babel blocks.

This commit removes "org-babel-default-header-args:es", and instead,
detects when :url and :method are unprovided in
"org-babel-execute:es". If either is missing, it is replaced with the
current value of the corresponding "es-default-*" variable.